refactored packaging and updated to version 0.1.7
