+++ logs/logfile.log
@@ -1,0 +1,23 @@
+-------------||time:2019-07-04T13:44:29.438909+0600||level:ERROR||-------------
+401 {'message': 'Bad credentials', 'documentation_url': 'https://developer.github.com/v3'}
+
+Traceback (most recent call last):
+  File "/home/redowan/anaconda3/envs/proto/bin/protomate", line 11, in <module>
+    load_entry_point('Protomate', 'console_scripts', 'protomate')()
+  File "/media/redowan/New Volume/protomate/protomate/script.py", line 174, in main
+    g, user = authentication(github_username, github_password)
+> File "/media/redowan/New Volume/protomate/protomate/script.py", line 85, in authentication
+    user.login
+  File "/home/redowan/anaconda3/envs/proto/lib/python3.7/site-packages/github/AuthenticatedUser.py", line 230, in login
+    self._completeIfNotSet(self._login)
+  File "/home/redowan/anaconda3/envs/proto/lib/python3.7/site-packages/github/GithubObject.py", line 263, in _completeIfNotSet
+    self._completeIfNeeded()
+  File "/home/redowan/anaconda3/envs/proto/lib/python3.7/site-packages/github/GithubObject.py", line 267, in _completeIfNeeded
+    self.__complete()
+  File "/home/redowan/anaconda3/envs/proto/lib/python3.7/site-packages/github/GithubObject.py", line 272, in __complete
+    self._url.value
+  File "/home/redowan/anaconda3/envs/proto/lib/python3.7/site-packages/github/Requester.py", line 275, in requestJsonAndCheck
+    return self.__check(*self.requestJson(verb, url, parameters, headers, input, self.__customConnection(url)))
+  File "/home/redowan/anaconda3/envs/proto/lib/python3.7/site-packages/github/Requester.py", line 286, in __check
+    raise self.__createException(status, responseHeaders, output)
+github.GithubException.BadCredentialsException: 401 {'message': 'Bad credentials', 'documentation_url': 'https://developer.github.com/v3'}
