+++ requirements.txt
@@ -1,8 +1,8 @@
+loguru==0.3.0
 pyfiglet==0.8.post1
-colorama==0.4.1
 setuptools==41.0.1
 termcolor==1.1.0
+colorama==0.4.1
 PyInquirer==1.0.3
 PyGithub==1.43.7
 prompt_toolkit==1.0.14
-loguru==0.3.0
